Switch to private repo for work config
If you are reading this, it's probably because someone (you?) discovered
that I had some IP addresses briefly stored here with work related
hosts. The IP addresses are either all in a private VPC, or in a few
cases, it looks like people used public EIPs that stuck around in config
but which were configured to require a special VPN for protection.

As far as I am aware, these are not secret IPs, they are intended to be
defended by networking constraints. IE, vpn or similar.