Fix csharp-ls support, again
Something is still preventing nixpkgs' chsarp-ls from working for me.

So instead, I've used mise to install dotnet@9 and then used that dotnet
to install csharp-ls while setting DOTNET_ROOT explicitly to allow it to
work. But then as an extra pain, I need to use dotnet@8 for the current
project... but this seems to work. Yikes! What a mess.